Configure racket/racket as primary repository
This commit changes the primary image repository to racket/racket, with jackfirth/racket preserved as a secondary repository for backwards compatibility. All images are pushed to both repositories, but the intent is that users of the jackfirth/racket repository should switch to the racket/racket repository. Both tags of the same images are tested in order to ensure they were tagged correctly and really do refer to the same image.

Also, this commit removes the DOCKER_USERNAME environment variable and replaces it with constants in _common.sh, since there's no particular reason that info needed to be passed through the CircleCI config. Note that the DOCKER_USERNAME variable, used to name the images, was distinct from the DOCKER_USER_NAME variable, which was used to login to DockerHub. The latter remains and is configured within CircleCI directly alongside an email and password.